more cleanup and simplification and planning
